Перенос ДБ контекста его изменение под универсальное подключение. Добавление Регистраторов под конкретные БД
