+++ Data/Thoughts.DAL.Entities/Role.cs
@@ -6,19 +6,20 @@ using System.Threading.Tasks;
 
 using Microsoft.EntityFrameworkCore;
 
-namespace Thoughts.DAL.Entities.Base
+using Thoughts.DAL.Entities.Base;
+
+namespace Thoughts.DAL.Entities;
+
+[Index(nameof(Name), IsUnique = true, Name = "NameIndex")]
+public class Role : Entity
 {
-    [Index(nameof(Name), IsUnique = true, Name = "NameIndex")]
-    public class Role : Entity
-    {
-        public string Name { get; set; } = null!;
+    public string Name { get; set; } = null!;
 
-        public ICollection<User> Users { get; set; } = new HashSet<User>();
-        public Role() { }
+    public ICollection<User> Users { get; set; } = new HashSet<User>();
+    public Role() { }
 
-        public Role(string name) => Name = name;
+    public Role(string name) => Name = name;
 
-        public override string ToString() => Name;
+    public override string ToString() => Name;
 
-    }
 }
