+++ Data/Thoughts.DAL.Entities/Status.cs
@@ -1,17 +1,18 @@
 ﻿
 using Microsoft.EntityFrameworkCore;
 
-namespace Thoughts.DAL.Entities.Base
+using Thoughts.DAL.Entities.Base;
+
+namespace Thoughts.DAL.Entities;
+
+[Index(nameof(Name), IsUnique = true, Name = "NameIndex")]
+public class Status:Entity
 {
-    [Index(nameof(Name), IsUnique = true, Name = "NameIndex")]
-    public class Status:Entity
-    {
-        /// <summary>Наименование статуса</summary>
-        public string Name { get; set; }= null!;
+    /// <summary>Наименование статуса</summary>
+    public string Name { get; set; }= null!;
 
-        public Status() { }
+    public Status() { }
 
-        public Status(string name) => Name = name;
-        public override string ToString() => Name;
-    }
+    public Status(string name) => Name = name;
+    public override string ToString() => Name;
 }
