+++ Data/Thoughts.DAL.Entities/Tag.cs
@@ -1,21 +1,22 @@
 ﻿using Microsoft.EntityFrameworkCore;
 
-namespace Thoughts.DAL.Entities.Base
+using Thoughts.DAL.Entities.Base;
+
+namespace Thoughts.DAL.Entities;
+
+/// <summary>
+/// Тег
+/// </summary>
+[Index(nameof(Name), IsUnique =true, Name = "NameIndex")]
+public class Tag:Entity
 {
-    /// <summary>
-    /// Тег
-    /// </summary>
-    [Index(nameof(Name), IsUnique =true, Name = "NameIndex")]
-    public class Tag:Entity
-    {
-        /// <summary>Название тега</summary>
-        public string Name { get; set; } = null!;
-        public ICollection<Post> Posts { get; set; }= new HashSet<Post>();
+    /// <summary>Название тега</summary>
+    public string Name { get; set; } = null!;
+    public ICollection<Post> Posts { get; set; }= new HashSet<Post>();
 
-        public Tag() { }
+    public Tag() { }
 
-        public Tag(string Name)=>this.Name = Name;
+    public Tag(string Name)=>this.Name = Name;
 
-        public override string ToString() => Name;
-    }
+    public override string ToString() => Name;
 }
