+++ Data/Thoughts.DAL.SqlServer/Thoughts.DAL.SqlServer.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="6.0.5" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Thoughts.DAL\Thoughts.DAL.csproj" />
+  </ItemGroup>
+
 </Project>
