+++ Data/Thoughts.DAL/Thoughts.DAL.csproj
@@ -14,4 +14,8 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Thoughts.DAL.Entities\Thoughts.DAL.Entities.csproj" />
+  </ItemGroup>
+
 </Project>
